Code cleanup, and update CSS
