+++ .changeset/clean-ghosts-grow.md
@@ -1,0 +1,5 @@
+---
+"expressive-code-twoslash": patch
+---
+
+More CSS cleanup, and organize code
