+++ docs/twoslash/src/content/docs/getting-started/installation.mdx
@@ -152,6 +152,7 @@ export default defineEcConfig({
 ecTwoSlash({
     explicitTrigger: true,
     includeJsDoc: true,
+    allowNonStandardJsDocTags: false,
     languages: ["ts", "tsx"],
     twoslashOptions: {},
 })
