+++ packages/twoslash/src/helpers/regex.ts
